Move settings into robot/common
